Make clientside switches consistently reflect actuator states
